==> 03_evaluate_model_performance.ipynb <==
#import relevant libraries
import sys; sys.path
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns
from statsmodels.stats.multitest import multipletests as fdr
import seaborn as sns

#load prediction accuracy data from optimised sex-independent and sex-specific models
fc_bf = pd.read_csv('corr_bf.txt', header=None).values
fc_bm = pd.read_csv('corr_bm.txt', header=None).values
fc_mf = pd.read_csv('corr_mf.txt', header=None).values
fc_mm = pd.read_csv('corr_mm.txt', header=None).values
fc_ff = pd.read_csv('corr_ff.txt', header=None).values
fc_fm = pd.read_csv('corr_fm.txt', header=None).values

#load prediction accuracy data from null models
null_fc_bf = np.median(pd.read_csv('corr_bf.txt', header=None).values, axis=0)
null_fc_bm = np.median(pd.read_csv('corr_bm.txt', header=None).values, axis=0)
null_fc_mf = np.median(pd.read_csv('corr_mf.txt', header=None).values, axis=0)
null_fc_mm = np.median(pd.read_csv('corr_mm.txt', header=None).values, axis=0)
null_fc_ff = np.median(pd.read_csv('corr_ff.txt', header=None).values, axis=0)
null_fc_fm = np.median(pd.read_csv('corr_fm.txt', header=None).values, axis=0)

#names of all the cognitive variables 
cognition = ['Crystal','Fluid','Total','PicVocab','Reading','Flanker','CardSort','PicSeq','ListSort','ProcSpeed']

#create a data frame for each cognitive variable with results from all the models 
crystal = pd.DataFrame({'bf' : fc_bf[:,0], 
                     'bm' : fc_bm[:,0], 
                     'mf' : fc_mf[:,0], 
                     'mm' : fc_mm[:,0], 
                     'ff' : fc_ff[:,0], 
                     'fm' : fc_fm[:,0]}) 

fluid = pd.DataFrame({'bf' : fc_bf[:,1], 
                     'bm' : fc_bm[:,1], 
                     'mf' : fc_mf[:,1], 
                     'mm' : fc_mm[:,1], 
                     'ff' : fc_ff[:,1], 
                     'fm' : fc_fm[:,1]}) 

total = pd.DataFrame({'bf' : fc_bf[:,2], 
                     'bm' : fc_bm[:,2], 
                     'mf' : fc_mf[:,2], 
                     'mm' : fc_mm[:,2], 
                     'ff' : fc_ff[:,2], 
                     'fm' : fc_fm[:,2]}) 

picvocab = pd.DataFrame({'bf' : fc_bf[:,3], 
                     'bm' : fc_bm[:,3], 
                     'mf' : fc_mf[:,3], 
                     'mm' : fc_mm[:,3], 
                     'ff' : fc_ff[:,3], 
                     'fm' : fc_fm[:,3]}) 

reading = pd.DataFrame({'bf' : fc_bf[:,4], 
                     'bm' : fc_bm[:,4], 
                     'mf' : fc_mf[:,4], 
                     'mm' : fc_mm[:,4], 
                     'ff' : fc_ff[:,4], 
                     'fm' : fc_fm[:,4]}) 

flanker = pd.DataFrame({'bf' : fc_bf[:,5], 
                     'bm' : fc_bm[:,5], 
                     'mf' : fc_mf[:,5], 
                     'mm' : fc_mm[:,5], 
                     'ff' : fc_ff[:,5], 
                     'fm' : fc_fm[:,5]}) 


cardsort = pd.DataFrame({'bf' : fc_bf[:,6], 
                     'bm' : fc_bm[:,6], 
                     'mf' : fc_mf[:,6], 
                     'mm' : fc_mm[:,6], 
                     'ff' : fc_ff[:,6], 
                     'fm' : fc_fm[:,6]}) 

picseq = pd.DataFrame({'bf' : fc_bf[:,7], 
                     'bm' : fc_bm[:,7], 
                     'mf' : fc_mf[:,7], 
                     'mm' : fc_mm[:,7], 
                     'ff' : fc_ff[:,7], 
                     'fm' : fc_fm[:,7]}) 

listsort = pd.DataFrame({'bf' : fc_bf[:,8], 
                     'bm' : fc_bm[:,8], 
                     'mf' : fc_mf[:,8], 
                     'mm' : fc_mm[:,8], 
                     'ff' : fc_ff[:,8], 
                     'fm' : fc_fm[:,8]}) 

procspeed = pd.DataFrame({'bf' : fc_bf[:,9], 
                     'bm' : fc_bm[:,9], 
                     'mf' : fc_mf[:,9], 
                     'mm' : fc_mm[:,9], 
                     'ff' : fc_ff[:,9], 
                     'fm' : fc_fm[:,9]}) 




#to evaluate whether each model is better than the null
def get_null_p(x,null):
    pval = 1-np.mean(x-null>=0)
    
    return pval

#to evaluate whether a pair of models are different in performance
def get_exact_p(x,y):
    pval = 2*np.min([np.mean(x-y>=0), np.mean(x-y<=0)])
    
    return pval

#evaluate predictions of crystallised abilities against null
crystal_p_fc_bf = get_null_p(fc_bf[:,0], null_fc_bf[0])
crystal_p_fc_bm = get_null_p(fc_bm[:,0], null_fc_bm[0])
crystal_p_fc_mf = get_null_p(fc_mf[:,0], null_fc_mf[0])
crystal_p_fc_mm = get_null_p(fc_mm[:,0], null_fc_mm[0])
crystal_p_fc_ff = get_null_p(fc_ff[:,0], null_fc_ff[0])
crystal_p_fc_fm = get_null_p(fc_fm[:,0], null_fc_fm[0])


crystal_pvals = [crystal_p_fc_bf, crystal_p_fc_bm, 
                 crystal_p_fc_mf, crystal_p_fc_mm,
                 crystal_p_fc_ff, crystal_p_fc_fm]

#evaluate predictions of fluid abilities against null
fluid_p_fc_bf = get_null_p(fc_bf[:,1], null_fc_bf[1])
fluid_p_fc_bm = get_null_p(fc_bm[:,1], null_fc_bm[1])
fluid_p_fc_mf = get_null_p(fc_mf[:,1], null_fc_mf[1])
fluid_p_fc_mm = get_null_p(fc_mm[:,1], null_fc_mm[1])
fluid_p_fc_ff = get_null_p(fc_ff[:,1], null_fc_ff[1])
fluid_p_fc_fm = get_null_p(fc_fm[:,1], null_fc_fm[1])


fluid_pvals = [fluid_p_fc_bf, fluid_p_fc_bm, 
                 fluid_p_fc_mf, fluid_p_fc_mm,
                 fluid_p_fc_ff, fluid_p_fc_fm]

#evaluate predictions of total abilities against null
total_p_fc_bf = get_null_p(fc_bf[:,2], null_fc_bf[2])
total_p_fc_bm = get_null_p(fc_bm[:,2], null_fc_bm[2])
total_p_fc_mf = get_null_p(fc_mf[:,2], null_fc_mf[2])
total_p_fc_mm = get_null_p(fc_mm[:,2], null_fc_mm[2])
total_p_fc_ff = get_null_p(fc_ff[:,2], null_fc_ff[2])
total_p_fc_fm = get_null_p(fc_fm[:,2], null_fc_fm[2])


total_pvals = [total_p_fc_bf, total_p_fc_bm, 
                 total_p_fc_mf, total_p_fc_mm,
                 total_p_fc_ff, total_p_fc_fm]

#evaluate predictions of pic vocab abilities against null
picvocab_p_fc_bf = get_null_p(fc_bf[:,3], null_fc_bf[3])
picvocab_p_fc_bm = get_null_p(fc_bm[:,3], null_fc_bm[3])
picvocab_p_fc_mf = get_null_p(fc_mf[:,3], null_fc_mf[3])
picvocab_p_fc_mm = get_null_p(fc_mm[:,3], null_fc_mm[3])
picvocab_p_fc_ff = get_null_p(fc_ff[:,3], null_fc_ff[3])
picvocab_p_fc_fm = get_null_p(fc_fm[:,3], null_fc_fm[3])


picvocab_pvals = [picvocab_p_fc_bf, picvocab_p_fc_bm, 
                 picvocab_p_fc_mf, picvocab_p_fc_mm,
                 picvocab_p_fc_ff, picvocab_p_fc_fm]

#evaluate predictions of reading abilities against null
reading_p_fc_bf = get_null_p(fc_bf[:,4], null_fc_bf[4])
reading_p_fc_bm = get_null_p(fc_bm[:,4], null_fc_bm[4])
reading_p_fc_mf = get_null_p(fc_mf[:,4], null_fc_mf[4])
reading_p_fc_mm = get_null_p(fc_mm[:,4], null_fc_mm[4])
reading_p_fc_ff = get_null_p(fc_ff[:,4], null_fc_ff[4])
reading_p_fc_fm = get_null_p(fc_fm[:,4], null_fc_fm[4])


reading_pvals = [reading_p_fc_bf, reading_p_fc_bm, 
                 reading_p_fc_mf, reading_p_fc_mm,
                 reading_p_fc_ff, reading_p_fc_fm]

#evaluate predictions of flanker abilities against null
flanker_p_fc_bf = get_null_p(fc_bf[:,5], null_fc_bf[5])
flanker_p_fc_bm = get_null_p(fc_bm[:,5], null_fc_bm[5])
flanker_p_fc_mf = get_null_p(fc_mf[:,5], null_fc_mf[5])
flanker_p_fc_mm = get_null_p(fc_mm[:,5], null_fc_mm[5])
flanker_p_fc_ff = get_null_p(fc_ff[:,5], null_fc_ff[5])
flanker_p_fc_fm = get_null_p(fc_fm[:,5], null_fc_fm[5])


flanker_pvals = [flanker_p_fc_bf, flanker_p_fc_bm, 
                 flanker_p_fc_mf, flanker_p_fc_mm,
                 flanker_p_fc_ff, flanker_p_fc_fm]

#evaluate predictions of cardsort abilities against null
cardsort_p_fc_bf = get_null_p(fc_bf[:,6], null_fc_bf[6])
cardsort_p_fc_bm = get_null_p(fc_bm[:,6], null_fc_bm[6])
cardsort_p_fc_mf = get_null_p(fc_mf[:,6], null_fc_mf[6])
cardsort_p_fc_mm = get_null_p(fc_mm[:,6], null_fc_mm[6])
cardsort_p_fc_ff = get_null_p(fc_ff[:,6], null_fc_ff[6])
cardsort_p_fc_fm = get_null_p(fc_fm[:,6], null_fc_fm[6])


cardsort_pvals = [cardsort_p_fc_bf, cardsort_p_fc_bm, 
                 cardsort_p_fc_mf, cardsort_p_fc_mm,
                 cardsort_p_fc_ff, cardsort_p_fc_fm]

#evaluate predictions of pic seq abilities against null
picseq_p_fc_bf = get_null_p(fc_bf[:,7], null_fc_bf[7])
picseq_p_fc_bm = get_null_p(fc_bm[:,7], null_fc_bm[7])
picseq_p_fc_mf = get_null_p(fc_mf[:,7], null_fc_mf[7])
picseq_p_fc_mm = get_null_p(fc_mm[:,7], null_fc_mm[7])
picseq_p_fc_ff = get_null_p(fc_ff[:,7], null_fc_ff[7])
picseq_p_fc_fm = get_null_p(fc_fm[:,7], null_fc_fm[7])


picseq_pvals = [picseq_p_fc_bf, picseq_p_fc_bm, 
                 picseq_p_fc_mf, picseq_p_fc_mm,
                 picseq_p_fc_ff, picseq_p_fc_fm]

#evaluate predictions of list sort abilities against null
listsort_p_fc_bf = get_null_p(fc_bf[:,8], null_fc_bf[8])
listsort_p_fc_bm = get_null_p(fc_bm[:,8], null_fc_bm[8])
listsort_p_fc_mf = get_null_p(fc_mf[:,8], null_fc_mf[8])
listsort_p_fc_mm = get_null_p(fc_mm[:,8], null_fc_mm[8])
listsort_p_fc_ff = get_null_p(fc_ff[:,8], null_fc_ff[8])
listsort_p_fc_fm = get_null_p(fc_fm[:,8], null_fc_fm[8])


listsort_pvals = [listsort_p_fc_bf, listsort_p_fc_bm, 
                 listsort_p_fc_mf, listsort_p_fc_mm,
                 listsort_p_fc_ff, listsort_p_fc_fm]

#evaluate predictions of proc speed abilities against null
procspeed_p_fc_bf = get_null_p(fc_bf[:,9], null_fc_bf[9])
procspeed_p_fc_bm = get_null_p(fc_bm[:,9], null_fc_bm[9])
procspeed_p_fc_mf = get_null_p(fc_mf[:,9], null_fc_mf[9])
procspeed_p_fc_mm = get_null_p(fc_mm[:,9], null_fc_mm[9])
procspeed_p_fc_ff = get_null_p(fc_ff[:,9], null_fc_ff[9])
procspeed_p_fc_fm = get_null_p(fc_fm[:,9], null_fc_fm[9])


procspeed_pvals = [procspeed_p_fc_bf, procspeed_p_fc_bm, 
                 procspeed_p_fc_mf, procspeed_p_fc_mm,
                 procspeed_p_fc_ff, procspeed_p_fc_fm]

#to highlight significant differences in tables
def color_negative_red(val):
    """
    Takes a scalar and returns a string with
    the css property `'color: red'` for negative
    strings, black otherwise.
    """
    color = 'red' if val > 0.05 else 'black'
    return 'color: %s' % color

#plot violin plots for all of the models
fig, axes = plt.subplots(2, 5, figsize=(20, 10), sharey=True, sharex=True)

#cognition = ['Crystal','Fluid','Total','PicVocab','Reading','Flanker','CardSort','PicSeq','ListSort','ProcSpeed']


sns.violinplot(ax=axes[0,0], data=crystal)
axes[0,0].set_title(cognition[0])

sns.violinplot(ax=axes[0,1], data=fluid)
axes[0,1].set_title(cognition[1])

sns.violinplot(ax=axes[0,2], data=total)
axes[0,2].set_title(cognition[2])

sns.violinplot(ax=axes[0,3], data=picvocab)
axes[0,3].set_title(cognition[3])

sns.violinplot(ax=axes[0,4], data=reading)
axes[0,4].set_title(cognition[4])

sns.violinplot(ax=axes[1,0], data=flanker)
axes[1,0].set_title(cognition[5])

sns.violinplot(ax=axes[1,1], data=cardsort)
axes[1,1].set_title(cognition[6])

sns.violinplot(ax=axes[1,2], data=picseq)
axes[1,2].set_title(cognition[7])

sns.violinplot(ax=axes[1,3], data=listsort)
axes[1,3].set_title(cognition[8])

sns.violinplot(ax=axes[1,4], data=procspeed)
axes[1,4].set_title(cognition[9])

axes[0,0].set_ylabel('prediction accuracy')

#concatenate all the pvals together for comparison against the null
pvals = np.concatenate((crystal_pvals, fluid_pvals, total_pvals, 
                        picvocab_pvals, reading_pvals,
                        flanker_pvals, cardsort_pvals, picseq_pvals, listsort_pvals, procspeed_pvals), axis=0)
#correct p-vals across all models for multiple comparisons
p_corr = fdr(pvals, alpha=0.05, method='fdr_bh', is_sorted=False, returnsorted=False)

#organise data into table
crystal_p_corr = p_corr[1][0:6]
fluid_p_corr = p_corr[1][6:12]
total_p_corr = p_corr[1][12:18]

picvocab_p_corr = p_corr[1][18:24]
reading_p_corr = p_corr[1][24:30]

flanker_p_corr = p_corr[1][30:36]
cardsort_p_corr = p_corr[1][36:42]
picseq_p_corr = p_corr[1][42:48]
listsort_p_corr = p_corr[1][48:54]
procspeed_p_corr = p_corr[1][54:60]

order = ['bf','bm','mf','mm','ff','fm']

model_null_eval = pd.DataFrame({'Model' : order, 
                                'Crystal' : crystal_p_corr, 
                                'Fluid' : fluid_p_corr,
                                'Total' : total_p_corr,
                                'Pic Vocab' : picvocab_p_corr, 
                                'Reading' : reading_p_corr,
                                'Flanker' : flanker_p_corr,
                                'Card Sort' : cardsort_p_corr, 
                                'Pic Seq' : picseq_p_corr,
                                'List Sort' : listsort_p_corr,
                                'Proc Speed' : procspeed_p_corr}) 

model_null_eval = model_null_eval.set_index('Model') 
model_null_eval_sig = model_null_eval.style.applymap(color_negative_red)

#show table where all models significantly better than null are shown in black, all other models in red
model_null_eval_sig

order = ['mm', 'mf', 'ff', 'fm']

for i in range(len(cognition)):
    
    #concatenate prediction accuracy results for sex-specific models
    corr =  [fc_mm[:,i], fc_mf[:,i],fc_ff[:,i], fc_fm[:,i]]
    p_model_comp = np.zeros([int(len(corr)),int(len(corr))])

    #check for differences usign exact test for differences
    for k in range(len(corr)):
         for j in range(len(corr)):
            if k==j:
                p_model_comp[k,j]=1
            elif k<j:
                p_model_comp[k,j]=1
            else:
                p_model_comp[k,j] = get_exact_p(corr[k],corr[j])

    #correct for multiple comparions
    p_corr = fdr(p_model_comp, alpha=0.05, method='fdr_bh', is_sorted=False, returnsorted=False)
    
    #organise into table
    model_comp = pd.DataFrame({'Model' : order, 
                           'mf' : p_model_comp[:,0],
                           'mm' : p_model_comp[:,1],
                           'ff' : p_model_comp[:,2],
                           'fm' : p_model_comp[:,3]})

    model_comp = model_comp.set_index('Model') 
    print(cognition[i])
    print(model_comp)
    print("")
