==> 04_feature_importance_reconstruction.ipynb <==
#import relevant libraries
import sys; sys.path
import pandas as pd
import numpy as np 
#import matplotlib.pyplot as plt
import seaborn as sns
from datetime import datetime

from sklearn.model_selection import train_test_split

#load in functional connectivity data + subj data file
fc = pd.read_csv('fc.csv', header=None).values
T = pd.read_csv('subj_data.csv',header=0)

#names of specific cognitive metrics you want to evaluate
cognition = ['Crystal','Fluid','Total','PicVocab','Reading','Flanker','CardSort','PicSeq','ListSort','ProcSpeed']


#extract the specific cognitive metrics
crystal = T.CogCrystalComp_AgeAdj.values
fluid = T.CogFluidComp_AgeAdj.values
total = T.CogTotalComp_AgeAdj.values
picvocab = T.PicVocab_AgeAdj.values
reading = T.ReadEng_AgeAdj.values
flanker = T.Flanker_AgeAdj.values
cardsort = T.CardSort_AgeAdj.values
picseq = T.PicSeq_AgeAdj.values
listsort = T.ListSort_AgeAdj.values
procspeed = T.ProcSpeed_AgeAdj.values

#put them all into one array
cog_metric = np.transpose(np.asarray([crystal, fluid, total, picvocab, reading, flanker, cardsort, picseq, listsort, procspeed]))

#set the # of cv loops and # of test loops and test size and number of cog variables to predict
perm = 100
cv_loops = 5
k = 3
train_size = .8
n_cog = np.size(cognition)

#set train data for males and females depending on what you want the input variable to be
X = fc

X_m = fc[T.Gender=='M',:]
Y_m = cog_metric[T.Gender=='M',:]

X_f = fc[T.Gender=='F',:]
Y_f = cog_metric[T.Gender=='F',:]

#set the number of features 
n_feat = X.shape[1]

featimp_m[:,:,0] = pd.read_csv('crystal_featimp_m.txt',header=None).values
featimp_f[:,:,0] = pd.read_csv('crystal_featimp_f.txt',header=None).values
featimp_b[:,:,0] = pd.read_csv('crystal_featimp_b.txt',header=None).values

featimp_m[:,:,1] = pd.read_csv('fluid_featimp_m.txt',header=None).values
featimp_f[:,:,1] = pd.read_csv('fluid_featimp_f.txt',header=None).values
featimp_b[:,:,1] = pd.read_csv('fluid_featimp_b.txt',header=None).values

featimp_m[:,:,2] = pd.read_csv('total_featimp_m.txt',header=None).values
featimp_f[:,:,2] = pd.read_csv('total_featimp_f.txt',header=None).values
featimp_b[:,:,2] = pd.read_csv('total_featimp_b.txt',header=None).values

featimp_m[:,:,3] = pd.read_csv('picvocab_featimp_m.txt',header=None).values
featimp_f[:,:,3] = pd.read_csv('picvocab_featimp_f.txt',header=None).values
featimp_b[:,:,3] = pd.read_csv('picvocab_featimp_b.txt',header=None).values

featimp_m[:,:,4] = pd.read_csv('reading_featimp_m.txt',header=None).values
featimp_f[:,:,4] = pd.read_csv('reading_featimp_f.txt',header=None).values
featimp_b[:,:,4] = pd.read_csv('reading_featimp_b.txt',header=None).values

featimp_m[:,:,5] = pd.read_csv('flanker_featimp_m.txt',header=None).values
featimp_f[:,:,5] = pd.read_csv('flanker_featimp_f.txt',header=None).values
featimp_b[:,:,5] = pd.read_csv('flanker_featimp_b.txt',header=None).values

featimp_m[:,:,6] = pd.read_csv('cardsort_featimp_m.txt',header=None).values
featimp_f[:,:,6] = pd.read_csv('cardsort_featimp_f.txt',header=None).values
featimp_b[:,:,6] = pd.read_csv('cardsort_featimp_b.txt',header=None).values

featimp_m[:,:,7] = pd.read_csv('picseq_featimp_m.txt',header=None).values
featimp_f[:,:,7] = pd.read_csv('picseq_featimp_f.txt',header=None).values
featimp_b[:,:,7] = pd.read_csv('picseq_featimp_b.txt',header=None).values

featimp_m[:,:,8] = pd.read_csv('listsort_featimp_m.txt',header=None).values
featimp_f[:,:,8] = pd.read_csv('listsort_featimp_f.txt',header=None).values
featimp_b[:,:,8] = pd.read_csv('listsort_featimp_b.txt',header=None).values

featimp_m[:,:,9] = pd.read_csv('procspeed_featimp_m.txt',header=None).values
featimp_f[:,:,9] = pd.read_csv('procspeed_featimp_f.txt',header=None).values
featimp_b[:,:,9] = pd.read_csv('procspeed_featimp_b.txt',header=None).values

#iterate through permutations
for p in range(perm):
    print('Permutation %d' %(p+1))

    cov_x_b = np.zeros([n_feat])
    cov_x_m = np.zeros([n_feat])
    cov_x_f = np.zeros([n_feat])
    
    x_train_m, x_test_m, cog_train_m, cog_test_m = train_test_split(X_m, Y_m, test_size=1-train_size, shuffle=True, random_state=p)
    x_train_f, x_test_f, cog_train_f, cog_test_f = train_test_split(X_f, Y_f, test_size=1-train_size, shuffle=True, random_state=p)

    
    x_train_b = np.concatenate((x_train_m, x_train_f), axis=0)
    cog_train_b = np.concatenate((cog_train_m, cog_train_f), axis=0)
    
    #compute covariance of training X data
    cov_x_b = np.cov(np.transpose(x_train_b))
    cov_x_m = np.cov(np.transpose(x_train_m))
    cov_x_f = np.cov(np.transpose(x_train_f))

    #for each of the cognitive scores
    for i in range(n_cog):
        #compute covariance of training cognitive scores
        cov_y_b = np.cov(cog_train_b[:,i])
        cov_y_m = np.cov(cog_train_m[:,i])
        cov_y_f = np.cov(cog_train_f[:,i])
        
        #compute activation vector by matrix multiplying training X covariance, feature weights, and 1/training Y covariance 
        A_b[p,:,i] = np.matmul(cov_x_b,featimp_b[p,:,i])*(1/cov_y_b)
        A_m[p,:,i] = np.matmul(cov_x_m,featimp_m[p,:,i])*(1/cov_y_m)
        A_f[p,:,i] = np.matmul(cov_x_f,featimp_f[p,:,i])*(1/cov_y_f)
        

#save data
np.savetxt('fc_392_coco439_crystal_A_m.txt', A_m[:,:,0], delimiter=',')
np.savetxt('fc_392_coco439_crystal_A_f.txt', A_f[:,:,0], delimiter=',')
np.savetxt('fc_392_coco439_fluid_A_m.txt', A_m[:,:,1], delimiter=',')
np.savetxt('fc_392_coco439_fluid_A_f.txt', A_f[:,:,1], delimiter=',')
np.savetxt('fc_392_coco439_total_A_m.txt', A_m[:,:,2], delimiter=',')
np.savetxt('fc_392_coco439_total_A_f.txt', A_f[:,:,2], delimiter=',')
np.savetxt('fc_392_coco439_picvocab_A_m.txt', A_m[:,:,3], delimiter=',')
np.savetxt('fc_392_coco439_picvocab_A_f.txt', A_f[:,:,3], delimiter=',')
np.savetxt('fc_392_coco439_reading_A_m.txt', A_m[:,:,4], delimiter=',')
np.savetxt('fc_392_coco439_reading_A_f.txt', A_f[:,:,4], delimiter=',')
np.savetxt('fc_392_coco439_flanker_A_m.txt', A_m[:,:,5], delimiter=',')
np.savetxt('fc_392_coco439_flanker_A_f.txt', A_f[:,:,5], delimiter=',')
np.savetxt('fc_392_coco439_cardsort_A_m.txt', A_m[:,:,6], delimiter=',')
np.savetxt('fc_392_coco439_cardsort_A_f.txt', A_f[:,:,6], delimiter=',')
np.savetxt('fc_392_coco439_picseq_A_m.txt', A_m[:,:,7], delimiter=',')
np.savetxt('fc_392_coco439_picseq_A_f.txt', A_f[:,:,7], delimiter=',')
np.savetxt('fc_392_coco439_listsort_A_m.txt', A_m[:,:,8], delimiter=',')
np.savetxt('fc_392_coco439_listsort_A_f.txt', A_f[:,:,8], delimiter=',')
np.savetxt('fc_392_coco439_procspeed_A_m.txt', A_m[:,:,9], delimiter=',')
np.savetxt('fc_392_coco439_procspeed_A_f.txt', A_f[:,:,9], delimiter=',')

np.savetxt('fc_392_coco439_crystal_A_b.txt', A_b[:,:,0], delimiter=',')
np.savetxt('fc_392_coco439_fluid_A_b.txt', A_b[:,:,1], delimiter=',')
np.savetxt('fc_392_coco439_total_A_b.txt', A_b[:,:,2], delimiter=',')
np.savetxt('fc_392_coco439_picvocab_A_b.txt', A_b[:,:,3], delimiter=',')
np.savetxt('fc_392_coco439_reading_A_b.txt', A_b[:,:,4], delimiter=',')
np.savetxt('fc_392_coco439_flanker_A_b.txt', A_b[:,:,5], delimiter=',')
np.savetxt('fc_392_coco439_cardsort_A_b.txt', A_b[:,:,6], delimiter=',')
np.savetxt('fc_392_coco439_picseq_A_b.txt', A_b[:,:,7], delimiter=',')
np.savetxt('fc_392_coco439_listsort_A_b.txt', A_b[:,:,8], delimiter=',')
np.savetxt('fc_392_coco439_procspeed_A_b.txt', A_b[:,:,9], delimiter=',')



#compute and save mean across 100 iterations
A_m_avg = np.mean(A_m,axis=0)
A_f_avg = np.mean(A_f,axis=0)
A_b_avg = np.mean(A_b,axis=0)


np.savetxt('fc_392_coco439_A_b_avg.txt', A_b_avg, delimiter=',')
np.savetxt('fc_392_coco439_A_m_avg.txt', A_m_avg, delimiter=',')
np.savetxt('fc_392_coco439_A_f_avg.txt', A_f_avg, delimiter=',')
